Noso makefile & nova estrutura
