+++ makefile
@@ -1,9 +1,14 @@
 DIR  = build
 NAME = omega
+FLAGS = -std=c17 -Wall -Wextra -Wpedantic -Iinclude
 
-compile: src/main.c src/$(NAME).c | $(DIR)
-	gcc -o $(DIR)/$(NAME) src/main.c src/$(NAME).c $(shell pkg-config --cflags --libs openssl)
-run: compile
-	./$(DIR)/$(NAME) Teste 3
+debug: src/main.c | $(DIR)
+	gcc $(FLAGS) -g -O0 src/*.c -o $(DIR)/debug/$(NAME)
+release: src/main.c | $(DIR)
+	gcc $(FLAGS) -O2 src/*.c -o $(DIR)/release/$(NAME)
+test: debug
+	./$(DIR)/debug/$(NAME) senhaTeste Chaveteste
 $(DIR):
-	@mkdir -p $(DIR)
+	@mkdir -p $(DIR)/debug
+	@mkdir -p $(DIR)/release
+	@mkdir -p $(DIR)/libs
